Added support for float values for experience stages and skill/magic rates.

+++ src/game.h
@@ -493,7 +493,7 @@ class Game
 		}
 
 		bool loadExperienceStages();
-		uint64_t getExperienceStage(uint32_t level);
+		float getExperienceStage(uint32_t level);
 
 		void loadMotdNum();
 		void saveMotdNum() const;
@@ -546,7 +546,7 @@ class Game
 		std::unordered_map<uint32_t, Player*> players;
 		std::unordered_map<std::string, Player*> mappedPlayerNames;
 		std::unordered_map<uint32_t, Guild*> guilds;
-		std::map<uint32_t, uint32_t> stages;
+		std::map<uint32_t, float> stages;
 
 		std::list<Item*> decayItems[EVENT_DECAY_BUCKETS];
 		std::list<Creature*> checkCreatureLists[EVENT_CREATURECOUNT];
